refactor: remove old not used code
