--- packages/solana/src/providers/index.ts
@@ -1,1 +1,0 @@
-export * from './WalletStandardProvider'
