+++ packages/solana/src/providers/shared/Errors.ts
